Итоговый коммит, всё готово!

+++ Allout/Allout/BusinessLogic.Core/Models/AuctionBlo.cs
@@ -6,8 +6,8 @@
         public int UserWhoUploadId { get; set; }
         public string LotName { get; set; }
         public string ImageUrl { get; set; }
-        public int StartCost { get; set; }
-        public int NowCost { get; set; }
+        public float StartCost { get; set; }
+        public float NowCost { get; set; }
         public bool IsDeleted { get; set; }
         public string Location { get; set; }
         public string Description { get; set; }
