+++ Allout/Allout/DataAccess.Core/Models/AuctionRto.cs
@@ -10,8 +10,8 @@ namespace Allout.DataAccess.Core.Models
         public UserRto UserWhoUpload { get; set; }
         public string LotName { get; set; }
         public string ImageUrl { get; set; }
-        public int StartCost { get; set; }
-        public int NowCost { get; set; }
+        public float StartCost { get; set; }
+        public float NowCost { get; set; }
         public string Location { get; set; }
         public bool IsDeleted { get; set; }
         public string Description { get; set; }
