+++ Allout/Allout/DataAccess.Core/Models/UserBalanceRto.cs
@@ -8,6 +8,6 @@ namespace Allout.DataAccess.Core.Models
         public int Id { get; set; }
         public int UserId { get; set; }
         public UserRto User { get; set; }
-        public int Balance { get; set; }
+        public float Balance { get; set; }
     }
 }
